# resources/platform-protocol-specs/worker-protocol-api.openapi.yaml
openapi: 3.1.0
info:
  title: durable-workflow.v2.worker-protocol-api
  version: "30"
  summary: Durable Workflow worker-plane HTTP+JSON API
  description: >
    Normative OpenAPI specification for the worker-plane HTTP+JSON API:
    worker registration, graceful worker deregistration, worker heartbeat,
    worker-session lifecycle,
    long-poll task acquisition, task history paging, task heartbeat,
    task completion, and task failure. Long-poll timing, cancellation,
    and lease semantics are specified by durable-workflow.v2.worker-protocol-stream.
    The server advertises its current 1.N version and accepts request headers
    from the same major whose minor is less than or equal to N.
...
servers:
  - url: /api
security:
  - durableWorkerAuth: []
paths:
...
  /worker/workflow-tasks/{taskId}/local-activities/checkpoint:
    post:
      operationId: checkpointLocalActivityPrefix
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        Commit the preceding command prefix using ordinary completion validation, authorization and quota. Retain the hosting claim. Checkpoint commit grants no callback admission.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalCheckpointRequest" }
      responses:
        "200":
          description: The durable checkpoint result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalCheckpointResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
  /worker/workflow-tasks/{taskId}/local-activities/checkpoint-group:
    post:
      operationId: checkpointLocalActivityGroup
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activity_groups
      description: >-
        Atomically commit every member of a complete all group, including prepare_local_activity descriptors without outcomes. Local executions are pending without attempts until separate preparation. Retain the original workflow claim, total deadlines and immutable group receipt. Partial and selection groups are refused.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalGroupCheckpointRequest" }
      responses:
        "200":
          description: The durable complete-group checkpoint result. Callback execution is not admitted.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalGroupCheckpointResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
...
  /worker/workflow-tasks/{taskId}/complete:
    post:
      operationId: completeWorkflowTask
      tags: [workflow-tasks]
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkflowTaskCompleteRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkflowTaskCompletion" }
        "409": { $ref: "#/components/responses/WorkflowTaskCompletionFailure" }
        "422": { $ref: "#/components/responses/WorkerError" }
...
  /worker/query-tasks/{queryTaskId}/complete:
    post:
      operationId: completeQueryTask
      tags: [query-tasks]
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - name: queryTaskId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueryTaskCompleteRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkerEnvelope" }
        "409": { $ref: "#/components/responses/WorkerError" }
...
components:
  securitySchemes:
    durableWorkerAuth:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    WorkerProtocolVersionHeader:
      name: X-Durable-Workflow-Protocol-Version
      in: header
      required: true
      description: >
        Worker-selected protocol version. The default server advertises 1.19
        and accepts same-major versions whose minor is no newer than 19.
        Missing, malformed, different-major, and ahead-of-server values fail closed.
      schema: { $ref: "#/components/schemas/AcceptedWorkerProtocolRequestVersion" }
    TaskIdPath:
      name: taskId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    WorkerEnvelope:
      description: Worker-protocol response envelope.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WorkerEnvelope" }
    WorkerError:
      description: Machine-readable worker-protocol error envelope.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WorkerError" }
    WorkerServiceUnavailable:
      description: >
        Temporary database connection loss returns backend_unavailable with an
        unknown operation outcome. Retry with bounded backoff and the identical
        worker, queue and poll_request_id; task=null does not prove that no task
        was claimed. Registration and worker heartbeat likewise do not acknowledge or
        reject an operation whose result is unknown. A workflow-task heartbeat
        carries task_id, lease_owner, and workflow_task_attempt so a worker can
        retry the same lease fence; a later 409 still revokes that lease. An
        activity completion carries task_id, activity_attempt_id and lease_owner
        to retry the same immutable result. Other
        503 errors retain their own contracts and are not automatically retryable. Optional
        storage admission returns storage_pressure or storage_admission_unavailable
        before admitting a request or a new poll claim. Retain the poll ID on
        retry: refusal of this attempt does not describe an earlier outcome.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema:
            anyOf:
              - $ref: "#/components/schemas/WorkerBackendUnavailable"
              - $ref: "#/components/schemas/WorkerStorageAdmissionRefusal"
              - $ref: "#/components/schemas/WorkerError"
...
    WorkflowTaskCompletion:
      description: A fenced workflow-task completion result.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WorkflowTaskCompletionResult" }
    WorkflowTaskCompletionFailure:
      description: A workflow-task completion boundary rejection.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
      content:
        application/json:
          schema:
            anyOf:
              - $ref: "#/components/schemas/WorkflowMetadataCapabilityFailure"
              - $ref: "#/components/schemas/ConditionWaitOccurrenceIdentityFailure"
              - $ref: "#/components/schemas/PortableWorkerAffinityFailure"
              - $ref: "#/components/schemas/WorkerError"
...
  schemas:
    AcceptedWorkerProtocolRequestVersion:
      type: string
      enum: ["1.0", "1.1", "1.2", "1.3", "1.4", "1.5", "1.6", "1.7", "1.8", "1.9", "1.10", "1.11", "1.12", "1.13", "1.14", "1.15", "1.16", "1.17", "1.18", "1.19"]
      description: >
        Request versions accepted by a server using the default advertised
        version 1.19. If deployment configuration changes the advertised
        version to 1.N, the accepted set is 1.0 through 1.N.
    AdvertisedWorkerProtocolVersion:
      type: string
      const: "1.19"
      description: Default worker protocol version advertised by responses implementing the Durable Workflow 2.0 protocol contract.
    WorkerEnvelope:
      type: object
      additionalProperties: true
      required: [protocol_version, server_capabilities]
      properties:
        protocol_version: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
        server_capabilities: { $ref: "#/components/schemas/WorkerServerCapabilities" }
    WorkerError:
      type: object
      additionalProperties: true
      required: [protocol_version, server_capabilities]
      properties:
        protocol_version: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
        error: { type: string }
        reason: { type: [string, "null"] }
        remediation: { type: [string, "null"] }
        server_capabilities: { $ref: "#/components/schemas/WorkerServerCapabilities" }
...
    WorkerServerCapabilities:
      type: object
      additionalProperties: true
      required: [supported_workflow_task_commands, workflow_task_poll_request_idempotency, workflow_memo_updates, message_streams, typed_search_attributes, condition_wait_occurrence_identity, update_validation_tasks, synchronous_update_validation, local_activities, sticky_execution]
      properties:
        cooperative_cancellation:
          type: boolean
          description: Explicit support for claim-bound cooperative request delivery, requiring worker protocol 1.20 or newer in the same major version. False at the default protocol 1.19.
        activity_cancellation_acknowledgement:
          type: boolean
          description: Original remote callback-stop receipts require an installed runtime primitive and worker protocol 1.20. False for the default protocol and for runtimes without that primitive.
        prepared_local_activities:
          type: boolean
          description: Candidate preparation, prefix checkpoint, supervisor control, application heartbeat, outcome, recovery and joined-stop receipts. Requires protocol 1.20 and the actual bound PreparedLocalActivityTaskBridge role including heartbeat support. False at the default protocol 1.19.
        prepared_local_activity_groups:
          type: boolean
          description: Candidate atomic complete-all-group admission. Requires prepared local activities, protocol 1.20 and the actual bound PreparedLocalActivityGroupTaskBridge. Existing custom prepared bridges do not acquire this role.
        supported_workflow_task_commands:
          type: array
          uniqueItems: true
          items: { type: string }
        workflow_task_poll_request_idempotency: { type: boolean, const: true }
        workflow_memo_updates:
          $ref: "#/components/schemas/WorkflowMemoUpdateCapability"
        message_streams:
          $ref: "#/components/schemas/MessageStreamsCapability"
        typed_search_attributes:
          $ref: "#/components/schemas/TypedSearchAttributesCapability"
        condition_wait_occurrence_identity:
          $ref: "#/components/schemas/ConditionWaitOccurrenceIdentityCapability"
        update_validation_tasks: { type: boolean, const: true }
        synchronous_update_validation:
          $ref: "#/components/schemas/SynchronousUpdateValidationCapability"
        worker_session_verbs:
          type: array
          items: { type: string, enum: [create, heartbeat, close] }
          minItems: 3
        worker_sessions:
          $ref: "./worker-sessions-runtime.schema.json#/$defs/workerSessionRuntimeContract"
        local_activities:
          $ref: "#/components/schemas/LocalActivityCapability"
        sticky_execution:
          $ref: "#/components/schemas/StickyExecutionCapability"
...
    PreparedLocalClaimRequest:
      type: object
      required: [lease_owner, workflow_task_attempt]
      properties:
        lease_owner: { type: string, minLength: 1, maxLength: 255 }
        workflow_task_attempt: { type: integer, minimum: 1 }
...
    PreparedLocalCheckpointRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalClaimRequest"
        - type: object
          required: [checkpoint_id, start_sequence, commands]
          properties:
            checkpoint_id: { type: string, minLength: 1, maxLength: 255 }
            start_sequence: { type: integer, minimum: 1 }
            commands:
              type: array
              maxItems: 100
              description: Nonterminal prefix commands. Turn-closing waits, local outcome reports and selection cancellation are excluded. Nonempty completion-only message cursor, wait and sticky cache metadata are rejected.
              items: { $ref: "#/components/schemas/WorkflowCommand" }
    PreparedLocalGroupCheckpointRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalCheckpointRequest"
        - type: object
          properties:
            commands:
              type: array
              minItems: 1
              maxItems: 100
              description: Complete all groups with at least one prepare_local_activity. Local descriptors use the PreparedLocalDescriptor fields with type prepare_local_activity and an authored parallel group path. They contain no outcome, result, routing or attempt reports. Cleanup proof is validated before any sibling is created.
              items: { $ref: "#/components/schemas/WorkflowCommand" }
...
    PreparedLocalCheckpointResult:
      type: object
      required: [checkpointed, duplicate, reason, checkpoint_id, task_id, workflow_run_id, workflow_task_attempt, start_sequence, next_sequence, fingerprint, created_task_ids, recorded_at]
      properties:
        checkpointed: { const: true }
        duplicate: { type: boolean }
        reason: { type: "null" }
        checkpoint_id: { type: string, minLength: 1 }
        task_id: { type: string, minLength: 1 }
        workflow_run_id: { type: string, minLength: 1 }
        workflow_task_attempt: { type: integer, minimum: 1 }
        start_sequence: { type: integer, minimum: 1 }
        next_sequence: { type: integer, minimum: 1 }
        fingerprint: { type: string, pattern: "^[0-9a-f]{64}$" }
        created_task_ids: { type: array, items: { type: string } }
        recorded_at: { type: string, format: date-time }
    PreparedLocalGroupCheckpointResult:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalCheckpointResult"
        - type: object
          required: [local_activities]
          properties:
            local_activities:
              type: array
              minItems: 1
              maxItems: 100
              items:
                type: object
                additionalProperties: false
                required: [sequence, activity_execution_id]
                properties:
                  sequence: { type: integer, minimum: 1 }
                  activity_execution_id: { type: string, minLength: 1 }
...
    WorkflowTaskCompleteRequest:
      type: object
      additionalProperties: true
      required: [lease_owner, workflow_task_attempt, commands]
      properties:
        lease_owner: { type: string, minLength: 1 }
        workflow_task_attempt: { type: integer, minimum: 1 }
        commands:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowCommand"
        message_stream_cursors:
          type: array
          maxItems: 100
          x-durable-workflow-minimum-protocol-version: "1.15"
          items: { $ref: "#/components/schemas/MessageStreamCursorAdvance" }
        message_stream_waits:
          type: array
          maxItems: 100
          x-durable-workflow-minimum-protocol-version: "1.15"
          items: { $ref: "#/components/schemas/MessageStreamWait" }
        sticky_cache:
          $ref: "#/components/schemas/StickyCacheClaim"
          x-durable-workflow-minimum-protocol-version: "1.18"
          x-durable-workflow-worker-capability: sticky_execution
...
    MessageStreamCursorAdvance:
      type: object
      additionalProperties: false
      required: [stream_name, through_position]
      properties:
        stream_name: { type: string, pattern: "^[A-Za-z0-9._:-]{1,128}$" }
        through_position: { type: integer, minimum: 0 }
...
    QueryTaskCompleteRequest:
      type: object
      additionalProperties: true
      required: [worker_id]
      properties:
        worker_id: { type: string, minLength: 1 }
        result: true
...